+++ package.json
@@ -13,6 +13,7 @@
     "@emotion/styled": "^10.0.27",
     "@popperjs/core": "^2.10.2",
     "@radix-ui/react-avatar": "^0.1.3",
+    "@radix-ui/react-collapsible": "^0.1.5",
     "@radix-ui/react-dialog": "^0.1.5",
     "@radix-ui/react-dropdown-menu": "^0.1.4",
     "@radix-ui/react-radio-group": "^0.1.4",
